Add missing "lib" prefix to native resolve logic
Move Any CPU native resolve logic to another file.

Minor changes to .editorconfig
